backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.36
geoalchemy2==0.15.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic-settings==2.0.3
httpx==0.27.0
shapely==2.0.6
geopandas==1.0.1
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
rasterio==1.3.10
pyproj==3.6.1
paho-mqtt==2.1.0
apscheduler==3.10.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
stripe==10.12.0
slowapi==0.1.9
